requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.31.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
alembic==1.14.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
ruff==0.14.6
pre-commit==4.5.0
black==25.11.0
fastapi-oauth2
authlib==1.6.5
pydantic[email]